move ManagedGroup model to separate file
