+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/model/api/SamJsonSupport.scala
@@ -11,9 +11,6 @@ import org.broadinstitute.dsde.workbench.sam.model.{
   FullyQualifiedPolicyId,
   FullyQualifiedResourceId,
   GroupSyncResponse,
-  ManagedGroupAccessInstructions,
-  ManagedGroupMembershipEntry,
-  ManagedGroupSupportSummary,
   OldTermsOfServiceDetails,
   PolicyIdentifiers,
   RequesterPaysSignedUrlRequest,
@@ -107,12 +104,6 @@ object SamJsonSupport {
   implicit val UserResourcesResponseFormat = jsonFormat6(UserResourcesResponse.apply)
 
   implicit val FullyQualifiedPolicyIdFormat = jsonFormat2(FullyQualifiedPolicyId.apply)
-
-  implicit val ManagedGroupMembershipEntryFormat = jsonFormat3(ManagedGroupMembershipEntry.apply)
-
-  implicit val ManagedGroupAccessInstructionsFormat = ValueObjectFormat(ManagedGroupAccessInstructions.apply)
-
-  implicit val ManagedGroupSupportSummaryFormat = jsonFormat4(ManagedGroupSupportSummary.apply)
 
   implicit val GroupSyncResponseFormat = jsonFormat2(GroupSyncResponse.apply)
